+++ isitfit/cli/cost.py
@@ -32,22 +32,20 @@ def analyze(ctx, filter_tags, save_details):
 
     share_email = ctx.obj.get('share_email', None)
 
-    from isitfit.cost.ec2.pipeline_factory import ec2_cost_analyze
-    from isitfit.cost.redshift.pipeline_factory import redshift_cost_analyze
+    # set up pipelines for ec2, redshift, and aggregator
+    from isitfit.cost import ec2_cost_analyze, redshift_cost_analyze, service_cost_analyze
     mm_eca = ec2_cost_analyze(ctx, filter_tags, save_details)
     mm_rca = redshift_cost_analyze(share_email, filter_region=ctx.obj['filter_region'], ctx=ctx)
 
-    # start download data and processing
-    logger.info("Fetching history: EC2...")
-    mm_eca.get_ifi()
-    logger.info("Fetching history: Redshift...")
-    mm_rca.get_ifi()
+    # combine the 2 pipelines
+    mm_all = service_cost_analyze(mm_eca, mm_rca, ctx, share_email)
 
-    # Display results
-    #logger.info("")
-    #logger.info("-"*20)
-    #logger.info("-"*20)
+    # configure tqdm
+    from isitfit.tqdmman import TqdmL2Account
+    tqdml2 = TqdmL2Account(ctx)
 
+    # Run pipeline
+    mm_all.get_ifi(tqdml2)