+++ isitfit/cost/__init__.py
@@ -1,0 +1,4 @@
+from isitfit.cost.ec2.pipeline_factory import ec2_cost_analyze
+from isitfit.cost.redshift.pipeline_factory import redshift_cost_analyze
+from isitfit.cost.service import service_cost_analyze
+